+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - App logo no redirect !27
 - Add `AllDefaultWeekdaysHaveAtLeastOneSession` constraint + fix all validation wording !28
 - Sort tutoring sessions by start date !29
+- Refactor subscribe js fetch flow with await !31
 - Changed multiple wording and added some notes !32
 
 ## [1.2.0] - 2023-09-25